rename to sg bazel and forward to bazel cli
